+++ frontend/src/pages/AdminAttendance/AdminAttendance.jsx
@@ -1,292 +1,194 @@
 import React, { useState } from 'react';
-import {
-  MainContentArea,
-  PageHeader,
-  PageTitle,
-  StatsCardGrid,
-  StatCard,
-  StatLabel,
-  StatValue,
-  StatUnit,
-  ControlPanel,
-  MonthNavigator,
-  NavButton,
-  CurrentMonth,
-  RefreshButton,
-  FilterTabs,
-  FilterTab,
-  TabIcon,
-  TabLabel,
-  TabBadge,
-  ScheduleTableWrapper,
-  ScheduleTable,
-  TableHeader,
-  TableHeaderCell,
-  TableBody,
-  TableRow,
-  TableCell,
-  FlightNumber,
-  RouteCode,
-  RouteArrow,
-  FlightTime,
-  Duration,
-  StatusBadge,
-  PassengerCount,
-  CrewSection,
-  CrewMember,
-  CrewRole,
-  CrewName,
-  CrewBadge,
-  ActionButton,
-} from './AdminAttendance.styled';
+import { Users, Clock, AlertCircle, CheckCircle } from 'lucide-react';
+import * as S from './AdminAttendance.styled';
 
 const AdminAttendance = () => {
-  // TODO: Zustand state mapping
-  // const { schedules, stats, fetchSchedules } = useScheduleStore();
-  
-  const [currentMonth, setCurrentMonth] = useState('2026년 1월');
-  const [activeTab, setActiveTab] = useState('assigned');
-  
-  // Mock data - Replace with Zustand
+  const [selectedTab, setSelectedTab] = useState('dashboard');
+
+  // Mock Data - 상단 통계
   const stats = [
-    { label: '전체 항공편', value: 100, unit: '편', color: '#e0e7ff' },
-    { label: '배정 완료', value: 100, unit: '편', color: '#d1fae5' },
-    { label: '수정 필요', value: 0, unit: '편', color: '#fef3c7' },
-    { label: '총 직원', value: 310, unit: '명', color: '#e5e7eb' },
+    { label: '전체 직원', value: 48, icon: Users, color: '#e0e7ff', textColor: '#4f46e5' },
+    { label: '출근', value: 42, icon: CheckCircle, color: '#d1fae5', textColor: '#059669' },
+    { label: '지각', value: 3, icon: Clock, color: '#fef3c7', textColor: '#d97706' },
+    { label: '결근', value: 3, icon: AlertCircle, color: '#fee2e2', textColor: '#dc2626' },
   ];
 
-  const filterTabs = [
-    { id: 'assigned', label: '배정 완료', icon: '✓', count: 100 },
-    { id: 'pending', label: '수정 필요', icon: '⚠', count: 0 },
-    { id: 'scheduled', label: '직원별 스케줄', icon: '📅', count: 190 },
+  // Mock Data - 휴가 승인 대기 목록
+  const [leaveRequests, setLeaveRequests] = useState([
+    {
+      id: 1,
+      name: '이수진',
+      department: '개발 1팀',
+      period: '2026-01-25 ~ 2026-01-27 (3일)',
+      reason: '개인 사유',
+      badge: '연차'
+    },
+    {
+      id: 2,
+      name: '이수진',
+      department: '개발 2팀',
+      period: '2026-01-25 ~ 2026-01-27 (3일)',
+      reason: '개인 사유',
+      badge: '연차'
+    },
+    {
+      id: 3,
+      name: '이수진',
+      department: '관리팀',
+      period: '2026-01-25 ~ 2026-01-27 (3일)',
+      reason: '개인 사유',
+      badge: '연차'
+    }
+  ]);
+
+  // Mock Data - 부서별 현황
+  const departmentStats = [
+    { name: '개발팀', total: 13, present: 13, late: 0, absent: 0, percentage: 100 },
+    { name: '디자인팀', total: 7, present: 7, late: 0, absent: 0, percentage: 100 },
+    { name: '기획팀', total: 9, present: 7, late: 1, absent: 1, percentage: 78 },
+    { name: '영업팀', total: 11, present: 9, late: 1, absent: 1, percentage: 82 },
+    { name: '인사팀', total: 2, present: 2, late: 0, absent: 0, percentage: 100 },
   ];
 
-  const scheduleData = [
-    {
-      id: 'KE101',
-      route: { from: 'ICN', to: 'AMS' },
-      departure: '01/15 06:15',
-      duration: '11.4h',
-      status: '완결',
-      passengers: 270,
-      crew: [
-        { role: '캡틴', name: '김영수 (CP)', badges: ['김인식', '이서연', '박지원', '+6명'] }
-      ]
-    },
-    {
-      id: 'KE102',
-      route: { from: 'ICN', to: 'NRT' },
-      departure: '01/15 07:30',
-      duration: '2h',
-      status: '단계별',
-      passengers: 250,
-      crew: [
-        { role: '캡틴', name: '이영호 (CP)', badges: ['원승익', '민서인', '오수빈', '+2명'] }
-      ]
-    },
-    {
-      id: 'KE103',
-      route: { from: 'ICN', to: 'BKK' },
-      departure: '01/15 08:45',
-      duration: '5.2h',
-      status: '단계별',
-      passengers: 290,
-      crew: [
-        { role: '캡틴', name: '박준형 (CP)', badges: ['최나은', '송지아', '백서준', '+3명'] }
-      ]
-    },
-    {
-      id: 'KE104',
-      route: { from: 'ICN', to: 'SIN' },
-      departure: '01/15 09:45',
-      duration: '6.8h',
-      status: '단계별',
-      passengers: 260,
-      crew: [
-        { role: '캡틴', name: '최동훈 (CP)', badges: ['안소희', '홍지현', '전민서', '+3명'] }
-      ]
-    },
-    {
-      id: 'KE105',
-      route: { from: 'ICN', to: 'FCO' },
-      departure: '01/15 10:00',
-      duration: '11.7h',
-      status: '완결',
-      passengers: 260,
-      crew: [
-        { role: '캡틴', name: '정승우 (CP)', badges: ['백수현', '양지혜', '우채원', '+6명'] }
-      ]
-    },
-    {
-      id: 'KE106',
-      route: { from: 'ICN', to: 'JFK' },
-      departure: '01/15 12:30',
-      duration: '13.7h',
-      status: '완결',
-      passengers: 280,
-      crew: [
-        { role: '캡틴', name: '김민수 (CP)', badges: ['곽준석', '공지명', '민서아', '+6명'] }
-      ]
-    },
-    {
-      id: 'KE107',
-      route: { from: 'ICN', to: 'SYD' },
-      departure: '01/15 13:15',
-      duration: '10.6h',
-      status: '완결',
-      passengers: 290,
-      crew: [
-        { role: '캡틴', name: '오재민 (CP)', badges: ['박민주', '표지현', '원소영', '+6명'] }
-      ]
-    },
-    {
-      id: 'KE108',
-      route: { from: 'ICN', to: 'FRA' },
-      departure: '01/15 14:45',
-      duration: '11h',
-      status: '완결',
-      passengers: 270,
-      crew: [
-        { role: '캡틴', name: '유성민 (CP)', badges: ['김소희', '나예은', '단지인', '+6명'] }
-      ]
-    },
-  ];
-
-  const handleMonthChange = (direction) => {
-    // TODO: Implement month navigation
-    console.log('Month navigation:', direction);
+  // 승인 처리
+  const handleApprove = (id) => {
+    console.log('승인:', id);
+    setLeaveRequests(prev => prev.filter(req => req.id !== id));
+    // TODO: API 호출
   };
 
-  const handleRefresh = () => {
-    // TODO: Implement data refresh with Zustand
-    console.log('Refreshing schedule data...');
-  };
-
-  const handleEdit = (scheduleId) => {
-    // TODO: Navigate to edit page
-    console.log('Edit schedule:', scheduleId);
+  // 반려 처리
+  const handleReject = (id) => {
+    console.log('반려:', id);
+    setLeaveRequests(prev => prev.filter(req => req.id !== id));
+    // TODO: API 호출
   };
 
   return (
-    <MainContentArea>
-      {/* Page Header */}
-      <PageHeader>
-        <div>
-          <PageTitle>관리자 근태 현황</PageTitle>
-        </div>
-      </PageHeader>
+    <S.Container>
+      {/* 헤더 */}
+      <S.Header>
+        <S.Title>근태관리 시스템</S.Title>
+        <S.UserInfo>관리자</S.UserInfo>
+      </S.Header>
 
-        {/* Statistics Cards */}
-        <StatsCardGrid>
-          {stats.map((stat, index) => (
-            <StatCard key={index} $color={stat.color}>
-              <StatLabel>{stat.label}</StatLabel>
-              <div>
-                <StatValue>{stat.value}</StatValue>
-                <StatUnit>{stat.unit}</StatUnit>
-              </div>
-            </StatCard>
-          ))}
-        </StatsCardGrid>
+      {/* 탭 네비게이션 */}
+      <S.TabContainer>
+        <S.Tab
+          $active={selectedTab === 'dashboard'}
+          onClick={() => setSelectedTab('dashboard')}
+        >
+          대시보드
+        </S.Tab>
+        <S.Tab
+          $active={selectedTab === 'attendance'}
+          onClick={() => setSelectedTab('attendance')}
+        >
+          직원 현황
+        </S.Tab>
+        <S.Tab
+          $active={selectedTab === 'calendar'}
+          onClick={() => setSelectedTab('calendar')}
+        >
+          근태 기록
+        </S.Tab>
+      </S.TabContainer>
 
-        {/* Control Panel */}
-        <ControlPanel>
-          <MonthNavigator>
-            <NavButton onClick={() => handleMonthChange('prev')}>
-              ‹
-            </NavButton>
-            <CurrentMonth>{currentMonth}</CurrentMonth>
-            <NavButton onClick={() => handleMonthChange('next')}>
-              ›
-            </NavButton>
-          </MonthNavigator>
+      {/* 메인 컨텐츠 */}
+      <S.MainContent>
+        {/* 상단 통계 카드 */}
+        <S.StatsGrid>
+          {stats.map((stat, index) => {
+            const IconComponent = stat.icon;
+            return (
+              <S.StatCard key={index} $color={stat.color}>
+                <S.StatIcon $color={stat.textColor}>
+                  <IconComponent size={24} />
+                </S.StatIcon>
+                <S.StatInfo>
+                  <S.StatLabel>{stat.label}</S.StatLabel>
+                  <S.StatValue $color={stat.textColor}>{stat.value}</S.StatValue>
+                </S.StatInfo>
+              </S.StatCard>
+            );
+          })}
+        </S.StatsGrid>
 
-          <RefreshButton onClick={handleRefresh}>
-            ↻ 자동 배정 실행
-          </RefreshButton>
-          <RefreshButton onClick={handleRefresh}>
-            ↻ 초기화
-          </RefreshButton>
-        </ControlPanel>
+        {/* 메인 레이아웃 (좌측: 휴가 승인 목록, 우측: 부서별 현황) */}
+        <S.ContentGrid>
+          {/* 좌측: 휴가 승인 대기 목록 */}
+          <S.LeftPanel>
+            <S.SectionHeader>
+              <S.SectionTitle>
+                <S.SectionIcon>📋</S.SectionIcon>
+                휴가 승인 대기 목록 (관리자)
+              </S.SectionTitle>
+              <S.FilterButtons>
+                <S.FilterButton $active>전체 보기</S.FilterButton>
+                <S.FilterButton>반려</S.FilterButton>
+              </S.FilterButtons>
+            </S.SectionHeader>
 
-        {/* Filter Tabs */}
-        <FilterTabs>
-          {filterTabs.map((tab) => (
-            <FilterTab
-              key={tab.id}
-              $active={activeTab === tab.id}
-              onClick={() => setActiveTab(tab.id)}
-            >
-              <TabIcon>{tab.icon}</TabIcon>
-              <TabLabel>{tab.label}</TabLabel>
-              <TabBadge $active={activeTab === tab.id}>
-                ({tab.count}건)
-              </TabBadge>
-            </FilterTab>
-          ))}
-        </FilterTabs>
-
-        {/* Schedule Table */}
-        <ScheduleTableWrapper>
-          <ScheduleTable>
-            <TableHeader>
-              <tr>
-                <TableHeaderCell>편명</TableHeaderCell>
-                <TableHeaderCell>노선</TableHeaderCell>
-                <TableHeaderCell>출발시간</TableHeaderCell>
-                <TableHeaderCell>비행시간</TableHeaderCell>
-                <TableHeaderCell>승객수</TableHeaderCell>
-                <TableHeaderCell>배정 승무원</TableHeaderCell>
-                <TableHeaderCell>관리</TableHeaderCell>
-              </tr>
-            </TableHeader>
-            <TableBody>
-              {scheduleData.map((schedule) => (
-                <TableRow key={schedule.id}>
-                  <TableCell>
-                    <FlightNumber>{schedule.id}</FlightNumber>
-                  </TableCell>
-                  <TableCell>
-                    <RouteCode>{schedule.route.from}</RouteCode>
-                    <RouteArrow>→</RouteArrow>
-                    <RouteCode>{schedule.route.to}</RouteCode>
-                  </TableCell>
-                  <TableCell>
-                    <FlightTime>{schedule.departure}</FlightTime>
-                  </TableCell>
-                  <TableCell>
-                    <Duration>{schedule.duration}</Duration>
-                    <StatusBadge $type={schedule.status}>
-                      {schedule.status}
-                    </StatusBadge>
-                  </TableCell>
-                  <TableCell>
-                    <PassengerCount>{schedule.passengers}명</PassengerCount>
-                  </TableCell>
-                  <TableCell>
-                    <CrewSection>
-                      {schedule.crew.map((member, idx) => (
-                        <CrewMember key={idx}>
-                          <CrewRole>{member.role}</CrewRole>
-                          <CrewName>{member.name}</CrewName>
-                          {member.badges.map((badge, bidx) => (
-                            <CrewBadge key={bidx}>{badge}</CrewBadge>
-                          ))}
-                        </CrewMember>
-                      ))}
-                    </CrewSection>
-                  </TableCell>
-                  <TableCell>
-                    <ActionButton onClick={() => handleEdit(schedule.id)}>
-                      ✏️ 수정
-                    </ActionButton>
-                  </TableCell>
-                </TableRow>
+            <S.LeaveRequestList>
+              {leaveRequests.map((request) => (
+                <S.LeaveRequestCard key={request.id}>
+                  <S.RequestBadge>{request.badge}</S.RequestBadge>
+                  <S.RequestInfo>
+                    <S.RequestName>{request.name}</S.RequestName>
+                    <S.RequestDepartment>{request.department}</S.RequestDepartment>
+                    <S.RequestPeriod>{request.period}</S.RequestPeriod>
+                    <S.RequestReason>사유: {request.reason}</S.RequestReason>
+                  </S.RequestInfo>
+                  <S.ActionButtons>
+                    <S.ApproveButton onClick={() => handleApprove(request.id)}>
+                      ✓ 승인
+                    </S.ApproveButton>
+                    <S.RejectButton onClick={() => handleReject(request.id)}>
+                      ✕ 반려
+                    </S.RejectButton>
+                  </S.ActionButtons>
+                </S.LeaveRequestCard>
               ))}
-            </TableBody>
-          </ScheduleTable>
-        </ScheduleTableWrapper>
-    </MainContentArea>
+            </S.LeaveRequestList>
+          </S.LeftPanel>
+
+          {/* 우측: 부서별 현황 */}
+          <S.RightPanel>
+            <S.SectionHeader>
+              <S.SectionTitle>
+                <S.SectionIcon>📊</S.SectionIcon>
+                부서별 현황
+              </S.SectionTitle>
+            </S.SectionHeader>
+
+            <S.DepartmentList>
+              {departmentStats.map((dept, index) => (
+                <S.DepartmentCard key={index}>
+                  <S.DepartmentHeader>
+                    <S.DepartmentName>{dept.name}</S.DepartmentName>
+                    <S.DepartmentTotal>{dept.total}/{dept.total}</S.DepartmentTotal>
+                  </S.DepartmentHeader>
+                  <S.ProgressBar>
+                    <S.ProgressFill $percentage={dept.percentage} />
+                  </S.ProgressBar>
+                  <S.DepartmentStats>
+                    <S.DepartmentStat $color="#059669">
+                      출근: {dept.present}
+                    </S.DepartmentStat>
+                    <S.DepartmentStat $color="#d97706">
+                      지각: {dept.late}
+                    </S.DepartmentStat>
+                    <S.DepartmentStat $color="#dc2626">
+                      결근: {dept.absent}
+                    </S.DepartmentStat>
+                  </S.DepartmentStats>
+                </S.DepartmentCard>
+              ))}
+            </S.DepartmentList>
+          </S.RightPanel>
+        </S.ContentGrid>
+      </S.MainContent>
+    </S.Container>
   );
 };
